backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]